requirements.txt
boto3
mysql-connector-python
requests
jsonschema
cfn_resource_provider
